Merge penguin:v2.5/linux into athlon.transmeta.com:/home/torvalds/v2.5/linux
2002/02/06 22:13:54-08:00 ak
[PATCH] Fix mount hash table

On my 512MB machine with 6 mount points the mount hash table uses 64K.
This patch brings it to a more reasonable size by limiting it to one
page.

2002/02/06 22:13:48-08:00 ak
[PATCH] Fix page cache limit wrapping in filesystems

Several file systems in tree that nominally support files >2GB set their
s_maxbytes value to ~0ULL. This has the nasty side effect on 32bit machines
that when a file write reaches the page cache limit (e.g. 2^43) it'll silently
wrap and destroy data at the beginning of the file.

This patch changes the file systems in question to fill in a proper limit.

I also have an alternate patch that adds a check for this generically
in super.c, but preliminary comments from Al suggested that he prefered
to do it in the file systems, so it is done this way way.

2002/02/06 22:13:41-08:00 ak
[PATCH] Automatic file-max sizing

The default for NR_FILES of 8192 is far too low for many workloads. This
patch does dynamic sizing for it instead. It assumes file+inode+dentry
are roughly 1K and will use upto 10% of the memory for it.

Also removes two obsolete prototypes.

2002/02/06 22:13:35-08:00 ak
[PATCH] Bootmem speedup

This patch speeds the bootmem freeing up a bit. Not particularly important,
but helps on some slow simulators where the loop can chew up significant
CPU time.

2002/02/06 22:13:29-08:00 ak
[PATCH] Slab name cleanup - last try

Hallo Linus,

This is the slab name cleanup you originally requested when the reiserfs
slab name too long boot bug was introduced. It replaces the static
name limit for slab cache names with the requirement for the caller
to supply a stable name. This version addresses all concerns that were given
over earlier versions of the patch.

As I do not feel very strongly about the slab cache name limit and you
also do not seem to be interested in it and the original bug of reiserfs
not booting anymore has been long fixed in your tree with the patch
you originally rejected from me I will not resend this patch after this
try.

Patch against 2.5.4pre1.

-Andi

BKrev: 3c621bb217lmcSD-_l_ErK2QlfKmLw