[PATCH] BUG register preservation
Here's the x86 BUG() implementation we discussed the other day.

I also have the rework of the header files which avoids
instantiation of strings in headers and saves 100-200k.
However that is only needed for gcc 2.X.  I assume that
by the time 2.6 is in use, gcc-3.x will be the preferred
compiler.

BKrev: 3c712f75Xtt70woLYOre4bGi-atUAQ