[PATCH] 2.5 dropping unneded lock precense check.
After BKL was moved from the VFS into the filesystem's methods in the lookup
call, we do not need to check BKL is actually held (we know it is not).
This patch patch removes unneeded lock_depth check.
Without this patch reiserfs cannot be used on SMP.

BKrev: 3c755632rqekFfH7_0VPu50JQv20cw