Merge penguin.transmeta.com:/home/penguin/torvalds/repositories/kernel/joystick into penguin.transmeta.com:/home/penguin/torvalds/repositories/kernel/linux
2002/02/26 13:26:19-08:00 vojtech
[PATCH] Input patch - joystick drivers update

As a step in the process of updating all the linux input drivers to the
most recent version available, this patch updates all the joystick
drivers that are in the kernel tree.

Changes are mainly the input hotplug/proc support in every of the
joystick drivers, and fixing trivial bugs here and there. I can supply a
more detailed description upon request.

BKrev: 3c7bfdack2QOXkgPk3V6_uGEnfr63g