[PATCH] PATCH 2/16: NFSD: BKL Removal: Lock export table
Change export table lock to (SMP safe) rwsemaphore

As a first step to removing the BKL from nfsd, this patch
changes the lock used for the export table to be a rwsem semaphore.
Previously it had the same functionality but depended on the BKL
for correctness.

As there is no "down_write_interruptible" this patch removes the
posibility of interrupting the write_lock request, but this should
never be needed anyway.

BKrev: 3c7c0aba7m0R3EC37ml55LFS1eTvKg