[PATCH] PATCH 5/16: NFSD: BKL removal: add BKL where needed in filehandle lookup
Protect dentry attachement from BKL

The process of attaching a dentry into the dcache
still needs the BKL I think.
When all the other BKL changes in the VFS settle down, I
will revisit this.  But as it is not a very frequent
operation, the BKL wont hurt.

Also add a down/up of i_sem when doing a lookup(".."),
as it is down for all other lookups.

BKrev: 3c7c0ac9JmyM7VeBWWy2x3XxkgyXPg