[PATCH] 2.5.6-pre1 IDE clean 13a
This is finally moving the ide-pci.c file into a shape where
the host chip detection lists can finally be moved to where they
belong - into the particular chipset specific files.

This is accomplished, by a rather obivous removal of macro magic,
which was just making entries to the global device type
list nonfunctional, instead of making them conditional on the
corresponding CONFIG_BLHA options.

The second thing was to add a flag field to the device recognition
list, which made it possible to compress many of the
multi || chip id conditionals go away.

The only other file affected is ide.h - here is the change in the size
of the name field, which apparently slipped through ide-clean-12...

BKrev: 3c7e8fe2rSVBrgAQ03SBH523Do3gnQ