[PATCH] SCSI cdrom cleanup
This is one of the very early steps on cleaning up the SCSI cdrom
driver.  It gets rid of directly accessing the scsi_CDs array in favour
of using the handle we get from the generic cdrom layer.  Also uses
local vars instead of many grouped scsi_CDs accesses in other places.

The gain is to get rid of the global, static array of CDROMS
in the end.

BKrev: 3c880a35ThriD8oELMUCTSe6DkmlMQ