[PATCH] PATCH - knfsd in 2.5.6 - fsid= export option
Support fsid=<number> export option to be device number independent

This patch was largely supplied by Steven Whitehouse <steve@gw.chygwyn.com>

A new export option "NFSEXP_FSID" indicates that the ex_dev passed down
is a user specified number, not a device number.
It should be used in fsid_type==1 filehandles to identify the
the exportpoint rather than the devid and inode (as in fsid_type == 0).
This allows filehandles to be device-number independent so that when Linux
changes device numbers on you (after reboot), your filesystems wont go stale.

User-space support for this is in the nfs-utils CVS and will be in
the next release (any release > 1.0).

BKrev: 3c8e6962zb_3kc_AqkyxjWs5xTPYHA