[PATCH] Trivial apm patch: move apm_error up
Here is the third one.

This patch moves the apm_error() function higher
in the file so that it is adjacent to the error_table
that it uses.  This makes the code easier to read.
The beginning of the file is an appropriate place
for "utility" functions of this kind.  This is a pure
move, with no changes made to the function.

BKrev: 3c8e6b75TUIwgr_8OmMzVW2a-vzk4w