[PATCH] PATCH 3 of 4 : knfsd : Store the fsid in the returned attributes instead of the device number
When a filesystem is exported with  fsid=  we should use that
fsid instead of the i_dev number when returning NFS attributes,
so that there is no chance of clients that depend on the filesys
id in the attributes getting confused by device numbers changing.

We only do this if the reference filehandle uses fsid to identify
the filesystem, so that a server can be converted from non-fsid= to
using fsid= without confusing active clients.

BKrev: 3cae2f21QTR6IcFeMe4optp8-ZIbbg