Merge home:v2.5/linux into penguin.transmeta.com:/home/penguin/torvalds/repositories/kernel/linux
2002/04/09 09:21:56-07:00 torvalds
Merge bk://linuxusb.bkbits.net/linus-2.5
into home.transmeta.com:/home/torvalds/v2.5/linux

2002/04/08 15:14:20-07:00 greg
USB

cleaned up the comments to put them in proper docbook format.

2002/04/08 15:01:34-07:00 david-b
This patch is a more complete fix for the device refcount
sanity checking and cleanup on device disconnect.

    - Splits apart usb_dec_dev_use(), for driver use, and
      usb_free_dev(), for hub/hcd use.  Both now have
      kerneldoc, and will BUG() if the refcount and the
      device tree get out of sync.  (Except for cleanup of
      root hub init errors, refcount must go to zero only
      at the instant disconnect processing completes.)

    - More usbcore-internal function declarations are
      now moved out of <linux/usb.h> into hcd.h

    - Driver-accessible refcounting is now inlined; minor
      code shrinkage, it's using atomic inc/dec instructions
      not function calls.

<note from greg k-h, there is still some work to be done with USB device
 reference counting, but this patch is a step in the right direction.>

2002/04/08 14:05:16-07:00 david-b
USB kerneldoc fixes

This fixes some kerneldoc bugs for USB.  It catches up with
the recent rename, and includes a couple minor tweaks/fixes
I happened to notice.

2002/04/08 11:25:54-07:00 greg
USB bluetooth tty driver

renamed the bluetooth.c file to bluettty.c to help prevent user confusion
between this driver and the bluez bluetooth stack in the kernel.

2002/04/08 10:51:56-07:00 ddstreet
[PATCH] usbfs disconnect

This was originally created by David many months ago and posted to the
list, but not put into the kernel.

I modified the original patch to:
-patch against the 2.5.7 kernel
-use the 'real' interface number, not position (to do this I added 2
 methods in usb.c)

2002/04/08 10:51:49-07:00 cph
[PATCH] page limit on bulk transfers in usbfs?

Here is an updated patch using kmalloc that removes the page limit
on bulk transfers in usbfs.  I've tested it and it seems to work OK.

2002/04/08 09:10:47-07:00 greg
changed direct access of current->state to set_current_state() call.

Thanks to Paul Komkoff for reminding me of this.

2002/04/08 09:05:05-07:00 johannes
[PATCH] uhci.c FSBR speed problem

The dangers of not ensuring important fields are properly set.

Some, if not all, of the speed problems with uhci.c were caused because
some important fields were not set and as a result, FSBR was never
turned on.

This patch also ensures is_suspended is set to 0 properly. This could
cause some problems suspending the HC when no devices are plugged in.

2002/04/08 09:04:56-07:00 johannes
[PATCH] uhci.c incorrect locking

This fixes a problem that I've not run into and is difficult to trigger,
but definately a bug. We locked urb->lock, when we meant u->lock. It also
cleans up the code a little to make it easier to understand and removes
an obsolete comment.

2002/04/08 09:04:48-07:00 johannes
[PATCH] uhci.c cleanup

This patch merely cleans up the code a little and doesn't fix any bugs.
It makes a couple of code paths a bit easier to understand, removes an
unused variable (uhci_list) and some procfs variables when not using
procfs.

2002/04/08 09:03:18-07:00 johannes
[PATCH] uhci.c SMP deadlock

This patch fixes a bug reported by Greg and David, as well as some other
people recently.

uhci.c would call wait_ms, which can sleep, with spinlocks held. There
are two places where this happens, but only one was the cause of the
problem.

2002/04/06 21:54:28-08:00 greg
USB

spilt up the Config.help file into smaller pieces

2002/04/06 20:17:46-08:00 david-b
USB usbnet driver update

    - adds ethtool support (based on code from Brad Hards)
    - makes diagnostic level configurable (ethtool, module param)
    - fixes a minor mem_flags goof (thanks Oliver!)
    - device identifers now use devpath (stable ID) not devnum

BKrev: 3cb3294f2ONmshdvJ8N7-1M9jYKpEA