[PATCH] ReiserFS get_block fix
This patch is to convert pap14030 panic into warning. While doing this,
a bug was uncovered, that when get_block() returns a failure, buffer
is still marked as mapped, and on subsequent access to this buffer
get_block() was not called anymore. This is also fixed.

BKrev: 3cb4a4086lU9_-XpuU4GyDdNIE86PA