[PATCH] cpqfc.o
This makes the cpqfc driver recognize the HP Tachyon.  I moved the
device list to an __initdata structure so the driver doesn't build it at
runtime and changed it to use the proper PCI_DEVICE_ID_* names.

With this patch applied, the driver happily detects the disks attached
to my HP Tachyon.

BKrev: 3cb9c968Okfa6cHHg-r7T30_4_2PyA